+++ packages/lint/eslint/eslint-config/index.js
@@ -1,4 +1,5 @@
 const { ifAnyDep } = require('@goldwasserexchange/read-pkg-up-helpers');
+const getOverridesFromLernaConfig = require('./getOverridesFromLerna');
 
 module.exports = {
   parser: 'babel-eslint',
@@ -27,4 +28,9 @@ module.exports = {
       jsx: ifAnyDep('react', true, false),
     },
   },
+  overrides: getOverridesFromLernaConfig({
+    react: ['@goldwasserexchange/eslint-config-react', 'eslint-config-airbnb/rules/react-a11y'],
+    ramda: '@goldwasserexchange/eslint-config-ramda',
+    jest: '@goldwasserexchange/eslint-config-jest',
+  }),
 };